refactor: move ecr tests into non-integrated since they only work tehre
