+++ containers/ecr-viewer/e2e/non-integrated/ecr-viewer.spec.ts
