+++ containers/ecr-viewer/e2e/non-integrated/utils.ts
